+++ examples/checkboxes/string_options/index.jsx
@@ -1,0 +1,33 @@
+import React, { useState } from 'react';
+import { Checkboxes } from '../../../src';
+
+const options = [
+  'Apple',
+  'Banana',
+  'Cherry',
+  'Mango',
+  'Ugli fruit',
+];
+
+export function Example() {
+  const [values, setValues] = useState(null);
+  return (
+    <>
+      <fieldset className="radios">
+        <legend>What is your favourite fruit</legend>
+        <Checkboxes
+          id="checkboxes"
+          values={values}
+          onValues={setValues}
+          options={options}
+        />
+      </fieldset>
+      <label htmlFor="output">
+        Current value
+      </label>
+      <output id="output">
+        {JSON.stringify(values, undefined, ' ')}
+      </output>
+    </>
+  );
+}