+++ src/components/checkboxes.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useNormalisedOptions } from '../hooks/use_normalised_options';
 import { renderGroupedOptions } from '../helpers/render_grouped_options';
@@ -6,18 +6,17 @@ import { classPrefix as defaultClassPrefix } from '../constants/class_prefix';
 import { makeBEMClass } from '../helpers/make_bem_class';
 import { visuallyHiddenClassName } from '../constants/visually_hidden_class_name';
 
-export function Radios(rawProps) {
+export function Checkboxes(rawProps) {
   const optionisedProps = Object.freeze({
     ...rawProps,
     ...useNormalisedOptions(rawProps),
   });
 
   const {
-    required,
     classPrefix,
     groupClassPrefix,
     name,
-    onValue,
+    onValues,
     options,
     renderWrapper,
     renderInput,
@@ -26,12 +25,40 @@ export function Radios(rawProps) {
     renderGroup,
     renderGroupLabel,
     renderGroupAccessibleLabel,
-    selectedOption,
+    selectedOptions,
   } = optionisedProps;
 
-  const handleChange = useCallback((e) => {
-    onValue?.(options.find((o) => o.identity === e.target.value)?.value);
-  }, [onValue, options]);
+  const checkboxRef = useRef(new Map());
+
+  const handleRef = useCallback((identity, el, ref) => {
+    if (el) {
+      checkboxRef.current.set(identity, el);
+    } else {
+      checkboxRef.current.delete(identity);
+    }
+
+    // Allow combination with a passed ref
+    if (typeof ref === 'function') {
+      ref(el);
+    } else if (ref) {
+      ref.current = el; // eslint-disable-line no-param-reassign
+    }
+  }, []);
+
+  function handleChange() {
+    // When getting the checked values,
+    // we need to get all the current checked values directly from the DOM
+    // If we only use the event target and combine with this selectedOptions
+    // we risk loosing values if selectedOptions is slow to update
+
+    const identities = [];
+    checkboxRef.current.forEach((value, key) => {
+      if (value.checked) {
+        identities.push(key);
+      }
+    });
+    onValues(options.filter((o) => identities.includes(o.identity)).map((o) => o.value));
+  }
 
   return renderGroupedOptions({
     options,
@@ -54,7 +81,7 @@ export function Radios(rawProps) {
     },
     renderOption(option) {
       const { identity, label, key, html, disabled, description, group } = option;
-      const checked = selectedOption?.identity === identity;
+      const checked = selectedOptions.some((item) => item.identity === identity);
 
       return renderWrapper({
         className: makeBEMClass(classPrefix),
@@ -67,11 +94,11 @@ export function Radios(rawProps) {
               id: key,
               name,
               onChange: handleChange,
-              type: 'radio',
+              type: 'checkbox',
               value: identity,
-              required,
               className: makeBEMClass(classPrefix, 'input'),
               ...html,
+              ref: (el) => handleRef(identity, el, html?.ref),
             }, { option, checked }, optionisedProps)}
             {renderLabel({
               htmlFor: key,
@@ -99,15 +126,13 @@ export function Radios(rawProps) {
   });
 }
 
-Radios.propTypes = {
+Checkboxes.propTypes = {
   classPrefix: PropTypes.string,
   groupClassPrefix: PropTypes.string,
   id: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  onValue: PropTypes.func,
+  onValues: PropTypes.func.isRequired,
   options: PropTypes.arrayOf(PropTypes.any).isRequired,
-  value: PropTypes.any,
-  required: PropTypes.bool,
+  values: PropTypes.arrayOf(PropTypes.any),
   renderWrapper: PropTypes.func,
   renderInput: PropTypes.func,
   renderLabel: PropTypes.func,
@@ -117,12 +142,10 @@ Radios.propTypes = {
   renderGroupAccessibleLabel: PropTypes.func,
 };
 
-Radios.defaultProps = {
-  classPrefix: `${defaultClassPrefix}radio`,
-  groupClassPrefix: `${defaultClassPrefix}radio-group`,
-  value: null,
-  onChange: null,
-  onValue: null,
+Checkboxes.defaultProps = {
+  classPrefix: `${defaultClassPrefix}checkbox`,
+  groupClassPrefix: `${defaultClassPrefix}checkbox-group`,
+  values: null,
   renderWrapper: (props) => <div {...props} />,
   renderInput: (props) => <input {...props} />,
   renderLabel: (props) => <label {...props} />,
@@ -130,5 +153,4 @@ Radios.defaultProps = {
   renderGroup: (props) => <div {...props} />,
   renderGroupLabel: (props) => <div {...props} />,
   renderGroupAccessibleLabel: (props) => <span {...props} />,
-  required: false,
 };