+++ src/components/index.js
@@ -4,4 +4,5 @@ export { DropDownTable } from './drop_down_table';
 export { ComboBox } from './combo_box';
 export { ComboBoxTable } from './combo_box_table';
 export { Radios } from './radios';
+export { Checkboxes } from './checkboxes';
 export * from './highlighters';